meta setup and remove mpir
